fix(daemon): split the entity surface registrations to clear the structural gates
PR #2215 went red on five structural facts, none of which touched the semantics
CEO already accepted:

- G0-5 shrink-only ratchet: the new `repo.entity.import` action pushed
  daemon-protocol-gui-actions.ts to 468 lines over its 456 baseline. The three
  entity-write ingresses (entity.import plus the agent/squad declaration writes)
  move into daemon-protocol-gui-actions-entity.ts and are spread back in at their
  original registry positions, so order and length both hold.
- file-complexity: entity-kind-registry.ts crossed 1100 lines. The bounded-context
  exception table is action-residency governance, not kind registration, so it
  moves to entity-bounded-context-exceptions.ts.
- schema-closure / derived-contracts: the three new reads declared an
  undeclared phase. They now carry Governed-Entity-W2-0, declared in the
  daemon protocol phase vocabulary beside the other GovernedEntity waves.
- The exhaustive GUI action facet table now registers repo.entity.import.
- The shipped-read sweep now supplies a locator payload, accepts the catalog
  envelope (which carries no `ok`), and asserts the three reads answer with
  real content through the resident daemon.
