+++ packages/daemon/test/json-rpc-protocol.test.ts
@@ -312,6 +312,7 @@ test("GUI action facets are exact, typed, and exclude the generic runner", () =>
     ["repo.decision.accept", { decisionId: "dec_A", rationale: "Approved", judgmentOnlyRationale: "Judgment" }],
     ["repo.decision.reject", { decisionId: "dec_A", reason: "Rejected" }],
     ["repo.decision.defer", { decisionId: "dec_A", reason: "Deferred" }],
+    ["repo.entity.import", { entityKind: "software/coding/architecture-decision-record@1", locator: "harness/adr/ADR-0001-example.md", expectedVersion: 0, title: "Example ADR" }],
     ["repo.receipt.show", { opId: "op_A" }],
     ["repo.settings.update", { defaultPreset: "strict-task", locale: "zh-CN", idempotencyKey: "settings-once" }],
     ["repo.gui.catalog.reread", {}],
